+++ frontend/src/scenes/persons/PersonPreview.tsx
@@ -1,13 +1,13 @@
 import { useValues } from 'kea'
 import { personLogic } from './personLogic'
 import { Spinner } from 'lib/lemon-ui/Spinner'
-import { asDisplay } from './PersonHeader'
 import { ProfilePicture } from 'lib/lemon-ui/ProfilePicture'
 import { LemonButton, Link } from '@posthog/lemon-ui'
 import { urls } from 'scenes/urls'
 import { PropertiesTable } from 'lib/components/PropertiesTable'
 import { PropertyDefinitionType } from '~/types'
 import { IconOpenInNew } from 'lib/lemon-ui/icons'
+import { asDisplay } from './person-utils'
 
 export type PersonPreviewProps = {
     distinctId: string | undefined
