+++ frontend/src/scenes/persons/personsLogic.tsx
@@ -16,12 +16,12 @@ import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { urls } from 'scenes/urls'
 import { teamLogic } from 'scenes/teamLogic'
 import { convertPropertyGroupToProperties, toParams } from 'lib/utils'
-import { asDisplay } from 'scenes/persons/PersonHeader'
 import { isValidPropertyFilter } from 'lib/components/PropertyFilters/utils'
 import { lemonToast } from 'lib/lemon-ui/lemonToast'
 import { TriggerExportProps } from 'lib/components/ExportButton/exporter'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { FEATURE_FLAGS } from 'lib/constants'
+import { asDisplay } from './person-utils'
 
 export interface PersonsLogicProps {
     cohort?: number | 'new'
